Added some automated tests.  (#46)
Added some automated tests.  These tests are rough and not thorough enough.  I am trying to cover things as broadly as possible.
Moved adding the Identity middleware to the router to each secured component.  This way the automated tests should tell us if we are missing he Identity middleware or not.